refactor: 커스텀훅으로 scrap 로직 분리
