+++ src/components/document/Basic.jsx
@@ -8,15 +8,15 @@ import DocumentLabel from "./DocumentLabel";
 import SelectInput from "@/components/common/input/SelectInput";
 import ScrapBtn from "@/components/common/button/ScrapBtn";
 import { basicModify } from "@/services/document";
-import { scrapCreate, scrapDelete } from "@/services/scrap";
 import "react-toastify/dist/ReactToastify.css";
 import { nullTokenEdit, adminApproval } from "@/utils/toast";
 import useDocsMutation from "@/hooks/useDocsMutation";
 import { ERROR_MSG, CATEGORY, DOCS_INFO } from "@/constant/document/create";
 import DocumentInputGroup from "@/components/createDocument/DocumentInputGroup";
+import useScrap from "@/hooks/useScrap";
 
 const Basic = ({ data }) => {
-  const { isLogin, memberId } = useSelector((state) => state.user);
+  const { isLogin } = useSelector((state) => state.user);
 
   const {
     id,
@@ -34,11 +34,9 @@ const Basic = ({ data }) => {
   let addressInfo = { lat: getLat, lng: getLng };
 
   const [isEditBasic, setIsEditBasic] = useState(false);
-  const [scraped, setScrap] = useState(isScraped);
 
   const { mutate: mutationBasicModify } = useDocsMutation(basicModify);
-  const { mutate: scrapDetailCreate } = useDocsMutation(scrapCreate);
-  const { mutate: scrapDetailDelete } = useDocsMutation(scrapDelete);
+  const { scraped, handleOnScrapFill } = useScrap(isScraped, id);
 
   const methods = useForm();
   const { handleSubmit, setValue, getValues, reset } = methods;
@@ -89,19 +87,6 @@ const Basic = ({ data }) => {
     setIsEditBasic(false);
     reset();
     setEditAddress(initialAddress);
-  };
-
-  useEffect(() => {
-    setScrap(isScraped);
-  }, [isScraped]);
-
-  const handleOnScrapFill = () => {
-    setScrap((prev) => !prev);
-    if (!scraped) {
-      scrapDetailCreate({ memberId, docsId: id });
-    } else {
-      scrapDetailDelete({ memberId, docsId: id });
-    }
   };
 
   return (
